#424 sorting still not what a human expects: needs to sort by *inferred* number, not just by string
