+++ js/components/preview/molecule/helperConstants.js
@@ -1,11 +1,3 @@
-const main_identifier = 'x';
-const sub_identifier = '_';
-
-export const constants = {
-  MAIN_IDENTIFIER: main_identifier,
-  SUB_IDENTIFIER: sub_identifier
-};
-
 export const moleculeProperty = {
   mw: 'MW',
   logP: 'logP',
