+++ js/components/preview/molecule/redux/selectors.js
@@ -1,43 +1,9 @@
 import { createSelector } from 'reselect';
-import { constants } from '../helperConstants';
 
 const getCachedMoleculeLists = state => state.apiReducers.cached_mol_lists;
 const getMoleculeGroupLists = state => state.apiReducers.mol_group_list;
 const getMoleculeGroupSelection = state => state.selectionReducers.mol_group_selection;
 const getObjectSelection = state => state.selectionReducers.object_selection;
-
-export function getNumberFromCode(inputCode) {
-  let number = 0;
-  let subNumber = 0;
-
-  let code = inputCode.toLowerCase();
-  let codeAfterIdentifier = code.split(constants.MAIN_IDENTIFIER)[1];
-
-  if (codeAfterIdentifier != null) {
-    let startingNumber = (codeAfterIdentifier.match(/\d+/) || [0])
-      .map(function(v) {
-        return +v;
-      })
-      .shift();
-
-    number = startingNumber;
-
-    if (codeAfterIdentifier.includes(constants.SUB_IDENTIFIER)) {
-      let codeAfterSubIdentifier = codeAfterIdentifier.split(constants.sub_identifier)[1];
-      if (codeAfterSubIdentifier != null) {
-        let startingSubNumber = (codeAfterSubIdentifier.match(/\d+/) || [0])
-          .map(function(v) {
-            return +v;
-          })
-          .shift();
-        subNumber = startingSubNumber;
-      }
-    }
-  }
-
-  let numberWithSub = +(number + '.' + subNumber);
-  return { number: numberWithSub };
-}
 
 export const selectJoinedMoleculeList = createSelector(
   getCachedMoleculeLists,
@@ -59,8 +25,7 @@ export const selectJoinedMoleculeList = createSelector(
           cachedDataArray = cachedData.results;
         }
         cachedDataArray.forEach(r => {
-          let result = getNumberFromCode(r.protein_code);
-          joinedMoleculeLists.push(Object.assign({ site: site, number: result.number }, r));
+          joinedMoleculeLists.push(Object.assign({ site: site }, r));
         });
       });
     }
