+++ js/constants/constants.js
@@ -1,3 +1,11 @@
+const main_identifier = 'x';
+const sub_identifier = '_';
+
+export const CONSTANTS = {
+  MAIN_IDENTIFIER: main_identifier,
+  SUB_IDENTIFIER: sub_identifier
+};
+
 export const VIEWS = {
   MAJOR_VIEW: 'major_view',
   SUMMARY_VIEW: 'summary_view',
